Quiet switch and improvement of console output.
Quiet mode to avoid printing to console. Coincidentally fixes the bug with piped output: attempting to set the cursor backwards caused an exception when the output was not the console.

Don't count down bytes remaining. Count up to 100% of (total) B instead. This obviates the need to move the cursor backwards. Only print when the percentage complete has changed by 0.1% - i.e. a total of 1000 updates.

more progress